YTDB-609: Fix integer overflow in guardOffset arithmetic
Widen pageOffset + accessWidth to long arithmetic to prevent silent
overflow when a stale read produces a pageOffset near Integer.MAX_VALUE.
Found during track-level code review gate check (C6).

+++ core/src/main/java/com/jetbrains/youtrackdb/internal/core/storage/impl/local/paginated/base/DurablePage.java
@@ -214,7 +214,7 @@ public class DurablePage {
    */
   private void guardOffset(final int pageOffset, final int accessWidth) {
     if (speculativeRead
-        && (pageOffset < 0 || pageOffset + accessWidth > buffer.capacity())) {
+        && (pageOffset < 0 || (long) pageOffset + accessWidth > buffer.capacity())) {
       throw OptimisticReadFailedException.INSTANCE;
     }
   }
